requirements.txt
py-clob-client>=0.29.0
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.28.0
